bot/requirements.txt
python-dotenv
python-telegram-bot
